rename image render file
