+++ src/db.json
@@ -1,0 +1,1 @@
+{"players": {"1": {"id": "account.15ea7ba0a57b4ed38d72f18a31f3f805", "name": "vhzZZZ", "lastMatchId": "", "analyzedMatches": ["cb5303e5-20e8-4ce7-be2d-04cc8dd6d0af"], "lastCheck": 1565434842.903687}}, "_default": {}, "authors": {"1": {"name": "glmn#4482", "id": 132402729887727616, "guild": 132403520102989824, "channelId": 604824085469462538, "players": ["account.15ea7ba0a57b4ed38d72f18a31f3f805"]}}, "guilds": {}}
